refactor: optimize daemon tests for performance and clarity
- Rename daemon_todo_test.go to daemon_features_test.go for clarity
- Eliminate real database connections in favor of mocking
- Add proper timeout handling to prevent test hangs
- Implement fast test mode with FAST_TESTS environment variable
- Fix race conditions in health check integration tests
- Add controlled logging output to reduce test noise
- Improve test execution time from 60s+ timeouts to <1s completion
